Simplify introduction on README
Features listed here are the same as on Shout repo, in same order, participating to the feeling of fork with nothing new. Instead of listing features here, we should refer to the website and improve it to make it as current as possible (there was some recent action there, and more coming, so it is reasonable to point there).

Also, this "Why the fork?" section was useful right when we forked, but now it gives unnecessary and lengthy information (it is now the most verbose section of the README!). The Lounge has enough momentum as that point to be treated as its own project.

Finally, shortening this section moves the screenshot back up on the page, and mobile view now has more context in the description.