Remove autocompletion section from the help page
Rationale for this is that the whole point of autocompletion is to be intuitive and show up naturally when starting to type something else. For example, all commands start with `/`, so obviously starting to type a command will trigger autocomplete. This is true for channels as well. Emoji are a bit particular because all systems that support emoji open their completion with `:`. The only not-so-intuitive completion strategy is for nicks because it is not so common to start them with `@` on IRC, but as long as we keep tab completion after any set of characters, this is fine. It will be even nicer once regular tab completion uses the same autocompletion dropdown.